==> 01- Exploration.ipynb <==
import pandas as pd
df_initial = pd.read_csv('feux_de_foret.csv')
df_initial.head()

df_initial.info()

df_initial.duplicated().sum()


df_initial.isna().sum()

#A number of variables will be useless for analysis, visualisation or the development of a prediction model. We will delete them in order to lighten the dataset.
#Most of the useful variables are complete. However, certain precise geographical coordinates (the name and FIPS code of the county) and the fire control date are often missing.


#This function is used to determine the number of NAs per variable and per State. The aim is to check that the NAs are correctly distributed.

STATE_LIST = sorted(list(df_initial.STATE.unique()))
df_NAs = pd.DataFrame()
df_NAs['STATES'] = STATE_LIST

def NA_Counters(colonne):
    
    col1 = str(colonne)+'.total'
    col2 = str(colonne)+'.percent'
    liste_total = []
    liste_percent = []
    
    for state in STATE_LIST:
       
        x = df_initial[df_initial['STATE'] == state][colonne].isna().sum()
        y = round((x/1880465)*100,2)
        liste_total.append(x)
        liste_percent.append(y)
        
    df_NAs[col1] = liste_total
    df_NAs[col2] = liste_percent
        
        
        

NA_Counters('DISCOVERY_TIME') 
NA_Counters('CONT_TIME') 
NA_Counters('FIPS_CODE') 
NA_Counters('FIPS_NAME') 

df_NAs.head()

#We note, for example, that California has high levels of NAs, even though it is one of the states most affected by fires.
#We store these results in a DataFrame so that they can be compared more easily.

df_NAs.sort_values(by='FIPS_CODE.percent', axis=0, ascending=False).head()
