==> 02- Data grouping and cleaning.ipynb <==
import seaborn as sns

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

# Data import

df = pd.read_csv('df_initial.csv')

df.head()

### Grouping by season

We'll create a season function to group lights on and off by season. We'll use the following mapping:

1: Summer: June/July/August
2: Autumn: September/October/November
3: Winter: December/January/February
4: Spring: March/April/May

def season(x):
    if 5 < x < 9:
        return '1'
    if 8 < x < 12:
        return '2'
    if x in [1,2,12]:
        return '3'
    if x in [3,4,5]:
        return '4'

Before applying the formula, we're going to transform the format of the discovery and extinction dates and extract the month and days of the week. Note that the variable 'CONT_DATE' contains 47% N/A. Next, we'll look at the duration of the fires and see if we can manage these N/As by using the statistics specific to the duration of the fires.

import datetime as datetime

#transform formats :

df['START_DATE'] = pd.to_datetime(df['DISCOVERY_DATE']- pd.Timestamp(0).to_julian_date(), unit='D')
df['END_DATE'] = pd.to_datetime(df['CONT_DATE'] - pd.Timestamp(0).to_julian_date(), unit='D')

df['START_MONTH'] = pd.DatetimeIndex(df['START_DATE']).month
df['START_DAY_OF_WEEK'] = pd.DatetimeIndex(df['START_DATE']).weekday

df['SEASON']=df['START_MONTH'].apply(season)

### Grouping by causes

We'll create a Human_Natural function to define whether the cause of the fire was natural, deliberate human or unintentional human. We will create a category to group causes that are not defined (~8% of the data). We will use the following mapping:

1: natural origin
2: voluntary human origin
3: involuntary human origin
0: unknown origin

df['STAT_CAUSE_DESCR'].value_counts(normalize=True)

def Human_Natural(x):
  if x in ['Missing/Undefined']:
    return 0
  if x in ['Lightning']:
    return 1
  if x in ['Arson']:
    return 2
  else:
    return 3

df['CAUSE_CATEG'] = df['STAT_CAUSE_DESCR'].apply(Human_Natural)

We will classify fires according to their size to define a Mega Fire class for fires that have burnt more than 24,000 acres (around 10,000 hectares). We will use the following mapping:

0: standard fires
1: mega fires

df['MEGA_FEUX'] = df['FIRE_SIZE'].apply(lambda x: 1 if x > 24710 else 0)

### Regroupement en régions

regions = []
North_East = ['CT','ME','MA','NH','VT','NJ','NY','RI','PA']
Midwest = ['MI','IL','IN','OH','WI','IA','KS','MN','MO','SD','ND','NE']
South = ['DE','FL','GA','MD','NC','SC','VA','DC','WV','AL','KY','MS','TN','AR','LA','TX','OK']
West = ['AZ','CO','MT','ID','NV','NM','WY','UT','CA','AK','HI','OR','WA']

divisions = []
New_England = ['CT','ME','MA','NH','VT','RI']
Mid_Atlantic = ['NJ','NY','PA']
East_North_Central = ['MI','IL','IN','OH','WI']
West_North_Central = ['IA','KS','MN','MO','SD','ND','NE']
South_Atlantic = ['DE','FL','GA','MD','NC','SC','VA','DC','WV']
East_South_Central = ['AL','KY','MS','TN']
West_South_Central = ['AR','LA','TX','OK']
Mountain = ['AZ','CO','MT','ID','NV','NM','WY','UT']
Pacific = ['CA','AK','HI','OR','WA']


def func_regions(series):
    for row in series:
        if row in North_East:
            regions.append('North_East')
        elif row in Midwest:
            regions.append('Midwest')
        elif row in South:
            regions.append('South')
        elif row in West:
            regions.append('West')
        else:
            regions.append('Other')

func_regions(df['STATE'])

def func_divisions(series):
    for row in series:
        if row in New_England:
            divisions.append('North East - New England')
        elif row in Mid_Atlantic:
            divisions.append('North East - Mid Atlantic')
        elif row in East_North_Central:
            divisions.append('Midwest - East North Central')
        elif row in West_North_Central:
            divisions.append('Midwest - West North Central')
        elif row in South_Atlantic:
            divisions.append('South- South Atlantic')
        elif row in East_South_Central:
            divisions.append('South - East South Central')
        elif row in West_South_Central:
            divisions.append('South - West South Central')
        elif row in Mountain:
            divisions.append('West - Mountain')
        elif row in Pacific:
            divisions.append('West - Pacific')
        else:
            divisions.append('Other')

func_divisions(df['STATE'])

df['REGIONS'] = regions
df['DIVISIONS'] = divisions

### Cleaned dataframe

In the next step, we're going to delete the columns we won't be using in the analysis.

to_drop = ['Unnamed: 0','STAT_CAUSE_CODE','STAT_CAUSE_DESCR','FOD_ID','FPA_ID','SOURCE_SYSTEM','SOURCE_SYSTEM_TYPE','NWCG_REPORTING_AGENCY','NWCG_REPORTING_UNIT_ID','NWCG_REPORTING_UNIT_NAME','SOURCE_REPORTING_UNIT','SOURCE_REPORTING_UNIT_NAME',
           'LOCAL_FIRE_REPORT_ID','LOCAL_INCIDENT_ID','FIRE_CODE','FIRE_NAME','ICS_209_INCIDENT_NUMBER','ICS_209_NAME','MTBS_ID','MTBS_FIRE_NAME','COMPLEX_NAME',
           'DISCOVERY_DOY','CONT_DOY','FIPS_CODE','FIPS_NAME','Shape','DISCOVERY_DATE','CONT_DATE','START_DATE','END_DATE','CONT_TIME','OWNER_CODE','OWNER_DESCR']
df_clean = df.drop(columns=to_drop)
df_clean.set_index(df_clean.OBJECTID,inplace=True)
df_clean.drop(columns='OBJECTID',inplace=True)
df_clean.head()

df_clean[df_clean['STATE']=='NY'].head()

df_clean['STATE'].value_counts()

df_clean.to_csv(r'C:\Users\norac\OneDrive\Documents\Reconversion\DataScientest\Projet Feux de Forêt\df_clean.csv',index=False)
