==> 03-  Visualization fires by state and year.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import plotly.express as px

df = pd.read_csv('feux_de_foret.csv',usecols=['STATE','FIRE_SIZE','FIRE_YEAR','FIRE_SIZE_CLASS','STAT_CAUSE_DESCR'])

regions = []
North_East = ['CT','ME','MA','NH','VT','NJ','NY','RI','PA']
Midwest = ['MI','IL','IN','OH','WI','IA','KS','MN','MO','SD','ND','NE']
South = ['DE','FL','GA','MD','NC','SC','VA','DC','WV','AL','KY','MS','TN','AR','LA','TX','OK']
West = ['AZ','CO','MT','ID','NV','NM','WY','UT','CA','AK','HI','OR','WA']


def func_regions(series):
    for row in series:
        if row in North_East:
            regions.append('North_East')
        elif row in Midwest:
            regions.append('Midwest')
        elif row in South:
            regions.append('South')
        elif row in West:
            regions.append('West')
        else:
            regions.append('Other')
            
func_regions(df['STATE'])



df['Regions'] = regions
#df['Divisions'] = divisions

categ = []

def functionF(series):
    for row in series:
        if row == 'Lightning':
            categ.append('Natural causes fire')
        elif row == 'Arson':
            categ.append('Criminal fire')
        else:
            categ.append('Unvolontary fire')

functionF(df['STAT_CAUSE_DESCR'])
df['STAT_CAUSE_DESCR'] = categ

df = df.rename(columns={'STAT_CAUSE_DESCR':'Fire origin'})
fig = px.histogram(df, x="Regions",color='Fire origin',color_discrete_sequence=['#335C81','#26C485','#d34e24'])
fig.update_layout(bargap=0.2)
fig.show()

figE = px.histogram(df,
                    'FIRE_YEAR',
                    color='Regions',
                    title='Nombre de feux par région et par an',
                    color_discrete_sequence = px.colors.sequential.YlOrRd,
                    facet_col='FIRE_SIZE_CLASS',
                    category_orders={"FIRE_SIZE_CLASS": ["A", "B", "C","D","E","F","G"]})

figE.update_layout(bargap=0.2,
                   xaxis_title = 'Année',
                   yaxis_title='Nombre de feux',
                   legend=dict(
                   orientation="h",
                   yanchor="bottom",
                   y=1.02,
                   xanchor="right",
                   x=1))

figE.for_each_annotation(lambda a: a.update(text=a.text.replace("FIRE_SIZE_CLASS=", "")))


fig = px.histogram(df,'FIRE_YEAR',color='Regions',title='Nombre de feux par région et par an',color_discrete_sequence = px.colors.sequential.YlOrRd)
fig.update_layout(bargap=0.2,xaxis_title = 'Année',yaxis_title='Nombre de feux',legend=dict(
    orientation="h",
    yanchor="bottom",
    y=1.02,
    xanchor="right",
    x=1
))
