==> 10- Prediction of the number and range of fires.ipynb <==
import seaborn as sns
import joblib

import pandas as pd
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

# Import

df = pd.read_csv('Dataframe_Final_Comtés_Mois.csv',index_col=0)

df.head()

df['Etat'] = df.Etat.astype('str')
df.info()

df['Etat'] = df['Etat'].apply(lambda x: 'State_' + x)
df.head()

#Creating dataframes for machine learning by state
def getdummies(df):
    state_county = pd.get_dummies(df['Etat_Comté'],prefix='State_County', prefix_sep='_')
    date = pd.get_dummies(df['Année'],prefix='Y', prefix_sep='_')
    df = df.join(date)
    df = df.join(state_county)
    df.drop(columns=['Année','Etat','Etat_Comté','Comté'],inplace=True)
    return df

liste = list(df.Etat.unique())
for state in liste:
    name = str(state)
    data = df[df['Etat'] == state]
    data = getdummies(data)
    data.to_csv(f'{name}.csv')

### Number of fires

#Creating a modelling function
def model(State):

    #Creating dataframe
    df_results = pd.DataFrame()
    nom = []
    Train_MSE = []
    Test_MSE = []
    Train_MAE = []
    Test_MAE = []

    name = str(State)
    df = pd.read_csv(f'{name}.csv',index_col=0)
    

    from sklearn.model_selection import train_test_split
    import sklearn.metrics
    from sklearn.metrics import mean_squared_error
    from sklearn.metrics import mean_absolute_error
    from sklearn.ensemble import RandomForestRegressor
    
    #target and explanatory variables + test and training sets
    features = df.drop(columns=['Etendue totale des feux','Nombre de feux'])
    nbre = df.loc[:,'Nombre de feux']
    etendue = df.loc[:,"Etendue totale des feux"]
    X_train,X_test,y_train,y_test = train_test_split(features,nbre,random_state=42,test_size=0.2)
    
    #model training and metrics
    regressor = RandomForestRegressor()
    regressor.fit(X_train,y_train)
    y_pred_test = regressor.predict(X_test)
    y_pred_test = pd.DataFrame(y_pred_test, columns=['predictions'])
    y_pred_train = regressor.predict(X_train)
    y_pred_train = pd.DataFrame(y_pred_train, columns=['predictions'])
    rfr_mse_test = mean_squared_error(y_test,y_pred_test)
    rfr_mae_test = mean_absolute_error(y_test,y_pred_test)
    rfr_mse_train = mean_squared_error(y_train,y_pred_train)
    rfr_mae_train = mean_absolute_error(y_train,y_pred_train)
    
    #retrieval for metrics
    nom.append(name)
    Train_MSE.append(rfr_mse_train)
    Train_MAE.append(rfr_mae_train)
    Test_MSE.append(rfr_mse_test)
    Test_MAE.append(rfr_mae_test)
    
    #dataset creation by state
    df_results['Etat']=nom
    df_results['Train_MSE']=Train_MSE
    df_results['Train_MAE']=Train_MAE
    df_results['Test_MSE']=Test_MSE
    df_results['Test_MAE']=Test_MAE
    
    #model export for streamlit
    joblib.dump(regressor, f'{name}'+'model.joblib')
    
    return df_results

#Application of the modelling function to States
df_resultats = pd.DataFrame()
for state in liste:
    df_resultats=pd.concat([df_resultats,model(state)], ignore_index=True)

df_resultats

df_resultats.max()

df_resultats.min()

df_resultats.mean()

### Fires' range

def model_etendue(State):

    #Creating dataframe
    df_results = pd.DataFrame()
    nom = []
    Train_MSE = []
    Test_MSE = []
    Train_MAE = []
    Test_MAE = []

    name = str(State)
    df = pd.read_csv(f'{name}.csv',index_col=0)
    from sklearn.model_selection import train_test_split
    import sklearn.metrics
    from sklearn.metrics import mean_squared_error
    from sklearn.metrics import mean_absolute_error
    
    features = df.drop(columns=['Etendue totale des feux','Nombre de feux'])
    nbre = df.loc[:,'Nombre de feux']
    etendue = df.loc[:,"Etendue totale des feux"]
    X_train,X_test,y_train,y_test = train_test_split(features,etendue,random_state=42,test_size=0.2)
    from sklearn.ensemble import RandomForestRegressor
    regressor = RandomForestRegressor()
    regressor.fit(X_train,y_train)
    y_pred_test = regressor.predict(X_test)
    y_pred_test = pd.DataFrame(y_pred_test, columns=['predictions'])
    y_pred_train = regressor.predict(X_train)
    y_pred_train = pd.DataFrame(y_pred_train, columns=['predictions'])
    rfr_mse_test = mean_squared_error(y_test,y_pred_test)
    rfr_mae_test = mean_absolute_error(y_test,y_pred_test)
    rfr_mse_train = mean_squared_error(y_train,y_pred_train)
    rfr_mae_train = mean_absolute_error(y_train,y_pred_train)
    
    nom.append(name)
    Train_MSE.append(rfr_mse_train)
    Train_MAE.append(rfr_mae_train)
    Test_MSE.append(rfr_mse_test)
    Test_MAE.append(rfr_mae_test)
    
    df_results['Etat']=nom
    df_results['Train_MSE']=Train_MSE
    df_results['Train_MAE']=Train_MAE
    df_results['Test_MSE']=Test_MSE
    df_results['Test_MAE']=Test_MAE
    return df_results

df_resultats_etendue = pd.DataFrame()
for state in liste:
    df_resultats_etendue=pd.concat([df_resultats_etendue,model_etendue(state)], ignore_index=True)
df_resultats_etendue
    

df_resultats_etendue.mean()

df_resultats_etendue.to_csv('df_resultats_etendue.csv')
df_resultats.to_csv('df_resultats_nombre.csv')
